+++ libs/blockchain_data_model/src/lib.rs
@@ -18,7 +18,7 @@ pub struct AssetPair {
 
 #[derive(Debug, PartialEq, SimpleSerialize, Default, Clone)]
 pub struct BlockFeedConfig {
-    pub id: u32,
+    pub id: u128,
     pub name: DataChunk,
     pub full_name: DataChunk,
     pub description: DataChunk,
@@ -71,7 +71,7 @@ impl BlockHeader {
 pub struct FeedActions {
     pub block_height: u64,
     pub new_feeds: [Option<BlockFeedConfig>; MAX_NEW_FEEDS_IN_BLOCK],
-    pub feed_ids_to_rm: [Option<u32>; MAX_FEED_ID_TO_DELETE_IN_BLOCK],
+    pub feed_ids_to_rm: [Option<u128>; MAX_FEED_ID_TO_DELETE_IN_BLOCK],
 }
 
 impl FeedActions {
